cephadm: move infer_local_ceph_image out of cephadm.py
Move infer_local_ceph_image out of cephadm.py into
cephadmlib.container_lookup continuing the overall effort to shrink
cephadm.py.
